% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-data-frame.R
\name{as.data.frame.insensitive}
\alias{as.data.frame.insensitive}
\title{Turn `httr::headers()`` objects into a data frame}
\usage{
\method{as.data.frame}{insensitive}(x, row.names = NULL,
  optional = FALSE, tbl = TRUE, ..., stringsAsFactors = FALSE)
}
\arguments{
\item{x}{an object returned from a call to \code{\link[httr:headers]{httr::headers()}}. This is
a \code{list} classed as \code{insensitive} which is pretty generic but
this method has a pretty thin use case and shouldn't cause issues.}

\item{row.names}{\code{NULL} or a character vector giving the row names for the
data frame. Missing values are not allowed.}

\item{optional}{ignored}

\item{tbl}{if \code{TRUE} (the default) will class the output as a tibble}

\item{...}{ignored}

\item{stringsAsFactors}{\emph{always} \code{FALSE} (unless you say otherwise)}
}
\description{
Turn `httr::headers()`` objects into a data frame
}
