% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrs.conf.R
\name{hdrs.conf}
\alias{hdrs.conf}
\title{Calculate bootstrap confidence intervals for the HDR-based synchrony metric}
\usage{
hdrs.conf(x, ci = 95, p = 80, nsim = 5000, nboot = 5000,
  ncores = 1)
}
\arguments{
\item{x}{A vector of individual completion times for the focal activity.}

\item{ci}{A number indicating the desired confidence level. The default is 95.}

\item{p}{A number indicating the percentage of individuals that have completed the focal activity. The default is 80.}

\item{nsim}{The number of simulations used to esimtate the HDR. The default is 5000.}

\item{nboot}{The number of bootstrap replicates used to estimate the CIs. The default is 5000.}

\item{ncores}{The number of processor cores over which to distribute the \code{nboot} replicates. The default is 1.}
}
\value{
A named numerical vector containing the lower and upper confidence limits for the \code{hdrs} metric.
}
\description{
Calculate bootstrap confidence intervals for the HDR-based synchrony metric
}
\note{
The \code{hdrs} metric quantifies the amount of time it took for \code{p} percent of the sampled individuals to
complete the specified activity (e.g., mating, calving, etc...). Therefore, a smaller \code{hdrs} value indicates a \emph{more}
synchronous population, while a larger \code{hdrs} value indicates \emph{less} synchrony.

The \code{ncores} argument specifies the number of processor cores over which the \code{foreach} loop
will distribute bootstrap re-samples. This should be set no higher than the number of hyperthreaded
cores on your machine, which is usually twice the number of physical cores.
}
\examples{
x <- rexp(25)
hdrs.conf(x, nboot=250)
}
