% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrs.test.R
\name{hdrs.test}
\alias{hdrs.test}
\title{A two-sample permutation test for differences in \code{hdrs} values.}
\usage{
hdrs.test(x, y, alternative = "two.tailed", p = 80, nsim = 5000,
  nperm = 5000, ncores = 1)
}
\arguments{
\item{x}{A vector of individual completion times for the focal activity.}

\item{y}{A vector of individual completion times for the focal activity.}

\item{alternative}{A character string specifying the alternative hypothesis, must be one of "two.sided" (default), "more.sync" or "less.sync".}

\item{p}{A number indicating the percentage of individuals that have completed the focal activity. The default is 80.}

\item{nsim}{The number of simulations used to esimtate the HDR. The default is 5000.}

\item{ncores}{The number of processor cores over which to distribute the \code{nboot} replicates. The default is 1.}

\item{ci}{A number indicating the desired confidence level. The default is 95.}

\item{nboot}{The number of bootstrap replicates used to estimate the CIs. The default is 5000.}
}
\value{
A list of class "hdrstest" containing the following values.
}
\description{
A two-sample permutation test for differences in \code{hdrs} values.
}
\note{
The \code{hdrs} metric quantifies the amount of time it took for \code{p} percent of the sampled individuals to
complete the specified activity (e.g., mating, calving, etc...). Therefore, a smaller \code{hdrs} value indicates a \emph{more}
synchronous population, while a larger \code{hdrs} value indicates \emph{less} synchrony. Consequently, the alternative
hypothesis that \code{x} is more synchronous than \code{y} ("more.sync") is equivalent to \code{hdrs(x) < hdrs(y)}, while
the "less.sync" alternative implies \code{hdrs(x) > hdrs(y)}. The null hypothesis is always \code{hdrs(x) = hdrs(y)}, and
the default "two.tailed" alternative is simply \code{hdrs(x) != hdrs(y)}.

The \code{ncores} argument specifies the number of processor cores over which the \code{foreach} loop
will distribute permutation re-samples. This should be set no higher than the number of hyperthreaded
cores on your machine, which is usually twice the number of physical cores.
}
\examples{
x <- rexp(25)
y <- rexp(25, rate=0.5)
hdrs.conf(x, y, alternative="more.sync", nperm=250)
}
