% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{http_headers}
\alias{http_headers}
\title{HTTP Header Reference}
\format{A data frame (classed as a tibble) with 184 rows and 14 observations}
\usage{
data(http_headers)
}
\description{
This dataset contains a very thorough reference on HTTP request and response
headers curated by the 'hsecsecan' project (\url{https://github.com/riramar/hsecscan}).
}
\details{
\itemize{
\item \code{header_field_name}: header field
\item \code{type_1}: \code{Permanent} (in a standard); \code{Provisional} (experimental); \code{Personal} (unofficial)
\item \code{protocol}: should always be \code{http} for now but may be different (e.g. \code{quic})
\item \code{status}: blank == unknown; otherwise the value describes the status well
\item \code{reference}: where to look for more info
\item \code{type_2}: \code{Request} (should only be found in requests); \code{Response} (should only be
found in responses); \code{Request/Response} found in either; \code{Reserved} (not in use yet)
\item \code{enable}: should you have this enabled
\item \code{required}: Is this header requried
\item \code{https}: HTTPS-specific header?
\item \code{security_description}: Information on the header
\item \code{security_reference}: Extra external reference information on the header
\item \code{recommendations}: Recommended setting(s)
\item \code{cwe}: Associated Common Weakness Enumeration (CWE) identifier
\item \code{cwe_url}: Associated CWE URL
}
}
\note{
Last updated 2019-03-05
}
\keyword{datasets}
