% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain_headers}
\alias{explain_headers}
\title{Explain HTTP headers found from a URL request}
\usage{
explain_headers(url, ...)
}
\arguments{
\item{url}{target URL}

\item{...}{passed on to the final \code{\link[httr:HEAD]{httr::HEAD()}} request.}
}
\value{
data frame (tibble)
}
\description{
Fetches headers from a URL and merges in information reference fields
from \link{http_headers}.
}
