% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrs-package.R
\docType{package}
\name{hdrs}
\alias{hdrs}
\alias{hdrs-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/hdrs}
\item BugReports: \url{https://gitlab.com/hrbrmstr/hdrs/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
