% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrs.R
\name{hdrs}
\alias{hdrs}
\title{Calculate the HDR-based synchrony metric}
\usage{
hdrs(x, p = 80, nsim = 5000)
}
\arguments{
\item{p}{A number indicating the percentage of individuals that have completed the focal activity. The default is 80.}

\item{nsim}{The number of simulations used to esimtate the HDR. The default is 5000.}
}
\value{
A number representing the length of time it takes \code{p} percent of indivdiuals to complete the focal activity.
}
\description{
Calculate the HDR-based synchrony metric
}
\note{
The \code{hdrs} metric quantifies the amount of time it took for \code{p} percent of the sampled individuals to
complete the specified activity (e.g., mating, calving, etc...). Therefore, a smaller \code{hdrs} value indicates a \emph{more}
synchronous population, while a larger \code{hdrs} value indicates \emph{less} synchrony.
}
\examples{
x <- rexp(25)
hdrs(x)
}
