% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrs.test.R
\name{hdrs.test}
\alias{hdrs.test}
\title{A two-sample permutation test for differences in \code{hdrs} values.}
\usage{
hdrs.test(x, y, alternative = "two.tailed", p = 80, nsim = 5000,
  nperm = 5000, ncores = 1)
}
\arguments{
\item{x}{A vector of individual completion times for the focal activity.}

\item{y}{A vector of individual completion times for the focal activity.}

\item{alternative}{A character string specifying the alternative hypothesis, must be one of "two.sided" (default), "more.sync" or "less.sync".}

\item{p}{A number indicating the percentage of individuals that have completed the focal activity. The default is 80.}

\item{nsim}{The number of simulations used to esimtate the HDR. The default is 5000.}

\item{ncores}{The number of processor cores over which to distribute the \code{nboot} replicates. The default is 1.}

\item{ci}{A number indicating the desired confidence level. The default is 95.}

\item{nboot}{The number of bootstrap replicates used to estimate the CIs. The default is 5000.}
}
\value{
A list of class "hdrstest" containing the following values.
}
\description{
A two-sample permutation test for differences in \code{hdrs} values.
}
\examples{
x <- rnorm(25)
y <- rnorm(25, sd=1.5)
hdrs.conf(x, y, alternative="more.sync")
}
